% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/livros.R
\name{get_capitulos_livros}
\alias{get_capitulos_livros}
\title{Extract book chapters}
\usage{
get_capitulos_livros(caminho_html, encoding = "ISO-8859-1")
}
\arguments{
\item{caminho_html}{Path to a Lattes HTML curriculum file.}

\item{encoding}{File encoding (default \code{"ISO-8859-1"}).}
}
\value{
A tibble with columns: autores, titulo_capitulo, titulo_livro,
organizadores, edicao, editora, cidade, pagina_inicial, pagina_final, ano,
id_lattes.
}
\description{
Extract book chapters
}
\examples{
html <- system.file("extdata", "exemplo.html", package = "getLattesHtml")
get_capitulos_livros(html)
}
