% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orientacoes.R
\name{get_orientacoes_mestrado}
\alias{get_orientacoes_mestrado}
\title{Extract master's advisorships}
\usage{
get_orientacoes_mestrado(caminho_html, encoding = "ISO-8859-1")
}
\arguments{
\item{caminho_html}{Path to a Lattes HTML curriculum file.}

\item{encoding}{File encoding (default \code{"ISO-8859-1"}).}
}
\value{
A tibble with columns: aluno, titulo, ano, curso, instituicao,
situacao, id_lattes.
}
\description{
Extract master's advisorships
}
\examples{
html <- system.file("extdata", "exemplo.html", package = "getLattesHtml")
get_orientacoes_mestrado(html)
}
