% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formacao.R
\name{get_formacao_graduacao}
\alias{get_formacao_graduacao}
\title{Extract undergraduate formation}
\usage{
get_formacao_graduacao(caminho_html, encoding = "ISO-8859-1")
}
\arguments{
\item{caminho_html}{Path to a Lattes HTML curriculum file.}

\item{encoding}{File encoding (default \code{"ISO-8859-1"}).}
}
\value{
A tibble with columns: periodo, curso, instituicao, titulo,
orientador, id_lattes.
}
\description{
Extract undergraduate formation
}
\examples{
html <- system.file("extdata", "exemplo.html", package = "getLattesHtml")
get_formacao_graduacao(html)
}
