% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/producao_tecnica.R
\name{get_patentes}
\alias{get_patentes}
\title{Extract patents and software registrations}
\usage{
get_patentes(caminho_html, encoding = "ISO-8859-1")
}
\arguments{
\item{caminho_html}{Path to a Lattes HTML curriculum file.}

\item{encoding}{File encoding (default \code{"ISO-8859-1"}).}
}
\value{
A tibble with columns: autores, titulo, ano, tipo, id_lattes.
}
\description{
Extract patents and software registrations
}
\examples{
html <- system.file("extdata", "exemplo.html", package = "getLattesHtml")
get_patentes(html)
}
